add update podcast job
